+++ app/models/dataPreprocessing/wordProcessor.py
@@ -1,7 +1,7 @@
 import nltk
 import string
 from nltk.corpus import stopwords
-from app.models.constants import Constants
+from app.static.constants.ModelConstants import ModelConstants
 
 """
 Remove common English stop words and punctuation from the input text.
@@ -43,7 +43,7 @@ def stop_word_removal(text: str) -> list:
 """
 
 
-def get_word_frequency_dist(words: list, count: int = Constants.COMMON_WORD_COUNT) -> list:
+def get_word_frequency_dist(words: list, count: int = ModelConstants.COMMON_WORD_COUNT) -> list:
     # Create a frequency distribution of the input list of words
     word_freq_dist = nltk.FreqDist(words)
