+++ app/static/constants/TextProcessingConstants.py
@@ -1,33 +1,17 @@
 import re
 
 """
-constants.py
-
-This module defines constants used throughout the project.
-
-Usage:
-- Import this module into your Python script or program to access the constants.
+   This class defines constants for various regular expressions used in text processing.
+   Each constant represents a specific pattern, such as URLs, hashtags, mentions, punctuation,
+   whitespaces, non-ASCII characters, and patterns for extracting skills, education, and experience sections.
 """
 
 
-class Constants:
-    """
-        This class defines general constants used in the project.
-    """
-
-    # Array group number
-    TEXT_GROUP = 1
-
-    # Common word count
-    COMMON_WORD_COUNT = 50
-
-
 class TextProcessingConstants:
-    """
-    This class defines constants for various regular expressions used in text processing.
-    Each constant represents a specific pattern, such as URLs, hashtags, mentions, punctuation,
-    whitespaces, non-ASCII characters, and patterns for extracting skills, education, and experience sections.
-    """
+
+    # Overrides the object method to restrict value modification.
+    def __setattr__(self, name, value):
+        raise AttributeError("Write Error: Cannot re-write constants values ")
 
     # Matches URLs
     URL_REGEX = 'http\S+\s*'
